Update to OpenTK trunk r1978.
This moves the e.g. OpenTK.Graphics.GLES11.* types into OpenTK.Graphics.ES11
(ditto for GLES10->ES10 and GLES20->ES20).  Aside from a namespace change,
this also re-cases many enumeration values, e.g. All.Colorbufferbit is now
All.ColorBufferBit (as it should be).

The GL types are now named ES.

The biggest change is to OpenTK.dll itself: OpenGLESSample only tests the ES11
types (ignoring everything else), and manually patching to add '#if IPHONE'
after every update is annoying, so instead of trying to include as much of
OpenTK as is possible, we're including as *little* of OpenTK as possible.

This amounts to the OpenTK.Graphics.ES*.* types and the
OpenTK.Platform.Utilities type.

I'll gradually move in the previous functionality once it's properly
refactored for minimal dependencies with upstream support (to minimize manual
changes).

OpenGLESSample is updated to cope with all of the above (mostly a s/GL/ES/g
operation).