Remove hasBeenCurrent.
The entire logic is faulty (and I'm not sure why the GLPaint demo even has it
in the first place).

The problem is that if the iPhoneOSGameView has CreateFrameBuffer() invoked
multiple times (as is the case in e.g. OpenGLESSample, as both
FinishedLaunching() and OnActivated() will cause it to be invoked), on the
second (and subsequent) invocation we'll call
gl.BindFramebuffer(All.FramebufferOes, oldFramebuffer), which is BAD as
'oldFramebuffer' is an INVALID framebuffer (as there was only ever one, and it
was destroyed before the "current" CreateFrameBuffer() invocation).

The result is annoying: if you override CreateFrameBuffer() to e.g. add a
depth buffer, the glFramebufferRenderbufferOES() call will fail, because the
framebuffer it's trying to attach to is the invalid one "restored" earlier!

For 1.0.x MonoTouch code, the workaround is to add the following code before
the GL.Oes.FramebufferRenderbuffer() call:

	GL.Oes.BindFramebuffer(All.FramebufferOes, base.Framebuffer);

After this change, the above line of code won't be necessary (but won't break
anything either).