Rework iPhoneOSGameView's idea of time. Should fix novell bug 686045 and assistly case 2687
DateTime's are bad when we want a monotonically increasing timer.  Using DateTime.Now
and DateTime.UtcNow both have problems (the latter has 1 fewer problems - it's immune
to timezone changes), Switch to using System.Diagnostics.Stopwatch for this instead.

Also refactor out all the "if (displayLink) else if (timer)" code from everywhere into
2 time sources.  The time source you get is chosen by whether you use Run() or
RunWithFrameInterval().

One side effect of this refactoring is that the DisplayLink property on iPhoneOSGameView
goes away, as it's an implementation detail of a particular time source.  This shouldn't
be an issue since it wasn't always present (if you ended up with the NSTimer implementation),
and exposing internal state is not something we should be doing anyway.