Don't call EglGetError() twice.
It appears that calling egl.EglGetError() clears the error value, such
that during the second call EGL10.EglSuccess is returned.

(At least that's my vain hope, because, the alternative will involve
pulling out more of my hair...)

For example, consider:
	http://lists.ximian.com/pipermail/monodroid/2012-May/010412.html

	"EglCreateWindowSurface failed with error 12288 (0x3000)"

0x3000 is EGL_SUCCESS. If the first call to egl.EglGetError() actually
returned EGL_SUCCESS, the message would be "EglCreateWindowSurface",
not "EglCreateWindowSurface failed...", so I think this bolsters the
theory that the error is cleared during the first EglGetError() call.