The AndroidGameView lockups up when Stop is called.
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=11921

The call to Wait on the render thread was causing a deadlock
with the SynchronizationContext. As a result the render
thread never exited. The solution is to just allow the
thread to exit normally without waiting and clean up the
CancellationTokenSource once the thread has exited.