[Bind] Add Android/iOS generation settings
In order to reduce the binary size, a new option “ForceDllImport” has
been added for iOS builds. This suppress the generation of entry point
slots, in favor of DllImports.

Android now uses “UseDllImport” by default. This suppress the
generation of entry point slots for core functions, and keeps slots
only for extension functions.