[Android] Corrected GetCurrentContext()
EGLContext.Handle returns the handle of the JNI object, not the handle
of the native EGL context. This means that different ContextHandles are
returned for the same EGL context.

According to the EGLContext documentation, GetHashCode() has the
expected behavior - the same hash code is always returned for the same
EGL context.