[OpenGLES] OpenTK app + Intent.CreateChooser() crashes
Fixes https://bugzilla.xamarin.com/show_bug.cgi?id=8182

Turns out the issue with this particular bug
was that we did not clear down the current context
prior to shutting down the activity/pausing. This left
OpenGL in some kind of weird state where it could not
use the surface and context current again until the
activity was destroyed.