Added Angle-based offscreen rendering.
AngleOffscreenRendering wraps a custom windowinfo and sets up the graphics context so a user can use OpenTK with an Angle backend.

The initialization is similar to GLControl, however, it should be platform independent (the platform is Angle, not Windows/Linux/Android....).

By default, we determine the backend automatically (D3D9 for ES2, D3D11 for ES3). Alternatively, a user can specify AngleD3D9, AngleD3D11 or AngleOpenGL explicitly.

Additional changes as part of this:

+ WindowInfo.CreatePbufferSurface was commented out. Enabled and fixed it.
+ GraphicsContextFlags.Offscreen will make the context create a PbufferSurface instead of a WindowSurface in EglContext