Added AngleWindowInfo and AnglePlatformFactory
Both WindowInfo and PlatformFactory wrap the actual platform-specific WindowInfo and PlatformFactory.

For offscreen rendering, AngleWindowInfo can be used with a DummyWindowInfo.

The API to add additional surfaces was added to IAngleWindowInfo, since all those calls need the Display parameter that is only available on EglWindowInfo (which is not exposed to users).