Fix bug where count overrides didn't effect ComputeSize
ComputeSize and ElementCount should match each other (and they do when
read direct from XML spec files). However when using the override file to
change count only ElementCount was effected. This is now fixed so that
ComputeSize is set and also so that ElementCount gets reset to 0 if the
count value isn't an integer.