Fixed incorrect Quaternion/Vector rotation
The issue was related to referencing...
In OpenTK 2.0, Vector3.Transform(Quaternion) actually returned a new instance of the rotated vector, which was then changed in 3.0 so that no new vector instances were created for better performance. In short, this is what caused the issue at Vector3.Transform(Quaternion), returning incorrectly rotated vectors.