Fix wrong binding code generation on some locales
The reason we set culture info (aka locale) here is because the binding
generator uses locale dependent string operations such as converting characters
to upper/lower case, case insensitive comparisons and float <-> string
conversions. This breaks in all kinds of nasty ways if you try to build the
project in parts of the world where upper case "version" is not "VERSION" but
rather "VERSİON" or lower case "VERSION" is 'versıon'. Obscure culture/language
conventions are among the things you will see that break software or make them
produce garbage output. This is a good examples of it and I suggest you study
it carefully. Theoretically this should be fixed by using overloads of those
string functions that take culture info parameters exclusively. This is
unfeasible however, more on that later.

You see, microsoft engineers thought that making all the relevant functions
such as float.Parse() locale dependent would make lives of programmers easier
since the functions would do the right thing™ by *default*. This is almost
NEVER what you want. Nobody in the right mind would ever think it is reasonable
for people write float.Parse(str, CultureInfo.InvariantCulture) every damn time
you don't want locale to be involved. It could've been an opt-in feature of the
library where you get locale dependent behavior if you _ask for it_ but no.
This shit is the reason your shitty csv/json parser crashes on a poor soul
living in some part of the world where comma instead of dot is used for decimal
numbers. This type of bullshit is the reason Visual Studio auto-complete used
to not show you "Invariant" when you enter "inv" because you happen to have a
tr-TR locale set on your computer. This right here is the reason some garbage
Asp.Net web page shows your purchase date in some retarded american format
because the server happens to be in USA.

Much can be said about how degenerate this global state dependent spooky action
at distance kind of nonsense is, to the point where you can convince anyone
that die-hard pure functional haskell fans were right all along. Things are
made worse by the fact that string interpolation and implicit string conversion
makes searching the problematic code rather inconceivable, and this cancerous
functionality is encouraged by the features integrated directly into the
language.

But the badness doesn't stop here. The people behind this glorious bullshit
were smart enough to make CultureInfo.CurrentCulture thread local (which btw
only prevents data races and retains all the disgusting ugliness a global state
can have), but not smart enough to make child threads inherit the locale from
its parent. Prior to .Net 4.5 every new thread would have its locale reset to
whatever the OS current is, which broke things even if you've set
CurrentCulture in the main thread but forgot to set it for each and every
thread you (or worse, a library you use) spawned!

At some point they must have realized how fucked up this whole situation is,
which resulted in an attempt to fix it by introducing _another_ property called
DefaultThreadCurrentCulture which determines what locale a new thread would get
initialized to. Don't be fooled into believing that this finally does solve
things sanely, as this is completely broken crap simply because it is an
application-wide global variable meaning that any thread at any time can change
it to facilitate subsequent threads they spawn to have a certain locale, except
of course it is not possible to do such thing in a multi-threaded environment
as you can not be sure if it changed just before you spawn a thread.
Unsurprisingly, the official docs for DefaultThreadCurrentCulture doesn't even
mention whether it is thread safe to mess with it as to keep you excited on
your coding journey.

So in order to avoid this utterly spectacular incompetence, we assign every
culture info setting to InvariantCulture and hope that no library under the sun
touches these abominations like CurrentCulture and its variants.