XI2MouseKeyboard: Process DeviceChangedEvent
Instead of querying every device when a DeviceChangedEvent is received,
process the event and update just the device that has changed.

This event is sent whenever a device changes its capabilities. This may
be sent when switching between two different mice or keyboards. Some
Wayland display servers send this very frequently, switching the mouse
between absolute and relative position mode whenever the mouse is
clicked or moved.

Processing the event instead of querying the devices fixes two issues:

  1. Because we are not resetting the mouse devices, we do not reset the
     mouse position.
  2. If there are very frequent DeviceChangedEvents, querying the
     devices makes us lose sync on the device's capabilities because the
     events are processed through a queue and querying is done
     synchronously. Mouse events received during this time may be
     processed incorrectly.